Update readme and use usize type

+++ std/Object.hh
@@ -3,7 +3,7 @@
 #include "prelude.h"
 
 class Object {
-    i32 refs = 1;
+    usize refs = 1;
 
     void init();
     virtual Object* ref();
